design / 텍스트 애니메이션, 이메일 입력창
